requirements.txt
alembic==1.14.0
babel==2.16.0
blinker==1.9.0
boto3==1.35.81
botocore==1.35.81
certifi==2024.12.14
click==8.1.7
cloudinary==1.41.0
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
flask-babel==4.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
Mako==1.3.8
MarkupSafe==3.0.2
packaging==24.2
pillow==11.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
s3transfer==0.10.4
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
WTForms==3.2.1